fix imagecache, prune .src, and don't touch files in .src

+++ src/scripts/dev/dev.backend.js
@@ -14,25 +14,31 @@ const srcDir = path.resolve('src')
 
 let backend = null
 
+const src = (src) => src.replace('/.src','/src')
+
 const dest = (src) => src.replace('/src','/.src')
 
-const getHashpath = (destpath) => {
-  const filename = destpath.replace(`${srcDir}/`, '').replace(`${srcDir}/`, '').split('/').join('.')+'.hash'
+const getHashpath = (src) => {
+  const filename = src.replace(/.*src\//,'').split('/').join('.')+'.hash'
   return path.join(dest(srcDir),'hash',filename)
 }
 
 const unlink = (src) => {
   const destpath = dest(src)
-  fs.unlinkSync(destpath)
+  const hashpath = getHashpath(src)
+  log('info', 'backend', `Unlinking ${destpath}`)
+  if(fs.existsSync(destpath)) fs.unlinkSync(destpath)
+  if(fs.existsSync(hashpath)) fs.unlinkSync(hashpath)
 }
 
-const write = (src, destpath, content) => {
+const write = (src, destpath, content, action) => {
   const hashpath = getHashpath(src)
   const hash = crypto.createHash('sha256').update(content, 'utf8').digest()
-  if(fs.existsSync(hashpath)) {
+  if(fs.existsSync(destpath) && fs.existsSync(hashpath)) {
     const oldhash = fs.readFileSync(hashpath, 'utf8')
     if(hash.toString('utf8')  === oldhash) return
   }
+  log('info', 'backend', `${action} ${src}`)
   fs.writeFileSync(destpath, content)
   fs.writeFileSync(hashpath, hash)
 }
@@ -41,7 +47,7 @@ const copy = (src) => {
   const destpath = dest(src)
   if(fs.existsSync(destpath) && fs.statSync(src).mtime <= fs.statSync(destpath).mtime) return
   const content = fs.readFileSync(src)
-  write(src, destpath, content)
+  write(src, destpath, content, 'Copying')
 }
 
 const transpile = (src) => {
@@ -49,12 +55,12 @@ const transpile = (src) => {
   if(fs.existsSync(destpath) && fs.statSync(src).mtime <= fs.statSync(destpath).mtime) return
   const source = fs.readFileSync(src, 'utf8')
   const transpiled = transform(source, babelrc)
-  write(src, destpath, transpiled.code)
+  write(src, destpath, transpiled.code, 'Compiling')
 }
 
-const buildFile = async (filepath) => {
+const buildFile = (filepath) => {
   if(path.extname(filepath) === '.js') return transpile(filepath)
-  await copy(filepath)
+  copy(filepath)
 }
 
 const buildItem = async (root) => {
@@ -64,7 +70,7 @@ const buildItem = async (root) => {
     const filepath = path.join(root, file)
     if(file.match(/_test.js$/)) return
     if(fs.lstatSync(filepath).isDirectory()) return await buildItem(filepath)
-    await buildFile(filepath)
+    buildFile(filepath)
   })
 }
 
@@ -75,9 +81,25 @@ const buildRoot = async () => {
   mkdirp.sync(path.join(destDir,'hash'))
   await Promise.map(fs.readdirSync(srcDir), async (file) => {
     const filepath = path.join(srcDir, file)
-    if(_.includes(['analytics','apps','core'], file)) return await buildItem(filepath)
-    if(path.extname(file) === '.js') await buildFile(filepath)
+    if(_.includes(['analytics','apps','core','public'], file)) return await buildItem(filepath)
+    if(path.extname(file) === '.js') buildFile(filepath)
   })
+}
+
+const cleanItem = async(root) => {
+  await Promise.map(fs.readdirSync(root), async (file) => {
+    const filepath = path.join(root, file)
+    const srcpath = src(filepath)
+    if(file === 'hash') return
+    if(fs.lstatSync(filepath).isDirectory()) return await cleanItem(filepath)
+    if(!fs.existsSync(srcpath)) unlink(filepath)
+  })
+}
+
+const cleanDotSrc = async (root) => {
+  log('info', 'backend', 'Cleaning dotsrc')
+  const destDir = dest(srcDir)
+  await cleanItem(destDir)
 }
 
 const watchSrc = async () => {
@@ -86,28 +108,25 @@ const watchSrc = async () => {
     path.resolve(srcDir,'*.js'),
     path.resolve(srcDir,'analytics'),
     path.resolve(srcDir,'apps'),
-    path.resolve(srcDir,'core')
+    path.resolve(srcDir,'core'),
+    path.resolve(srcDir,'public')
   ], {
     ignoreInitial: true,
     ignored: (absolute) => {
       const pathname = absolute.replace(srcDir,'src')
       if(pathname.match(/web/)) return true
       if(pathname.match(/admin\/(activities|badges|components|roots|tokens|views)/)) return true
-      if(pathname.match(/_test.js$/)) return true
+      if(pathname.match(/_test\.js$/)) return true
       return false
     }
   }).on('all', async (event, absolute) => {
-    const pathname = absolute.replace(srcDir,'src')
     if(event === 'add') {
-      log('info', 'backend', `Adding ${pathname}`)
-      return await buildFile(absolute)
+      return buildFile(absolute)
     }
     if(event === 'change') {
-      log('info', 'backend', `Recompiling ${pathname}`)
-      return await buildFile(absolute)
+      return buildFile(absolute)
     }
     if(event === 'unlink') {
-      log('info', 'backend', `Removing ${pathname}`)
       return unlink(absolute)
     }
   })
@@ -121,17 +140,17 @@ const start = () => {
       NODE_ENV: 'development'
     }
   })
-  proc.stdout.on('data', function (data) {
+  proc.stdout.on('data', (data) => {
     process.stdout.write(data)
   })
-  proc.stderr.on('data', function (data) {
+  proc.stderr.on('data', (data) => {
     process.stdout.write(data)
   })
   return proc
 }
 
 const restartBackend = _.debounce(() => {
-  log('info', 'backend', 'Restarting dotsrc')
+  log('info', 'backend', 'Restarting backend')
   backend.kill('sighup')
   backend = start()
 }, 500)
@@ -142,13 +161,23 @@ const watchDotSrc = async () => {
   chokidar.watch(dest(srcDir), {
     ignoreInitial: true
   }).on('all', (event, absolute) => {
-    log('info', 'backend', `${event} ${absolute}`)
+    const srcpath = src(absolute)
+    const srcExists = fs.existsSync(srcpath)
+    if(event === 'add' && !srcExists) {
+      log('error', 'backend', 'Dont touch files in .src')
+      return unlink(absolute)
+    }
+    if(event === 'unlink' && srcExists) {
+      log('error', 'backend', 'Dont touch files in .src')
+      return buildFile(srcpath)
+    }
     restartBackend()
   })
 }
 
 const watchBackend = async () => {
   await buildRoot()
+  await cleanDotSrc()
   await watchSrc()
   await watchDotSrc()
 }
